+++ .eslintrc.yml
@@ -54,3 +54,4 @@ rules:
   semi: [2, always]
   linebreak-style: [1, unix]
   no-plusplus: 2
+  react/jsx-filename-extension: [1, { "extensions": [".js", ".jsx"] }]
