+++ src/components/CurrentBalance/components/Balance/styles.js
