+++ src/components/Modules/Accounts/Goals/index.jsx
@@ -10,18 +10,10 @@ import {
 } from './styles';
 
 const Control = ({ title }) => {
-  const [goals, setGoals] = useState([]);
-  const [displayGoals, setDiplayGoals] = useState([]);
-  const [totalGoals, setTotalGoals] = useState({});
-
   // Hooks
   useEffect(() => {
     const getGoals = async () => {
       const { display, items, total } = await mock.getGoals();
-
-      setGoals(items);
-      setDiplayGoals(display);
-      setTotalGoals(total);
     };
     getGoals();
   }, []);
