+++ src/index.js
@@ -4,6 +4,6 @@ import ReactDOM from 'react-dom';
 import App from './App';
 
 ReactDOM.render(
-    <App />,
-  document.getElementById('root')
+  <App />,
+  document.getElementById('root'),
 );
