+++ src/pages/Accounts/index.jsx
@@ -2,14 +2,21 @@ import React from 'react';
 
 import Header from '../../components/Header';
 import CurrentBalance from '../../components/CurrentBalance';
+import Goals from '../../components/Modules/Accounts/Goals';
+import BillsToPay from '../../components/Modules/Accounts/BillsToPay';
+import Division from '../../components/Division';
 
-import Module from '../../modules/Accounts';
+import { Module } from './styles';
 
 const Accounts = () => (
   <>
     <Header title="Financial Targets" />
     <CurrentBalance />
-    <Module />
+    <Module>
+      <Division title="Contas" />
+      <Goals title="Metas" />
+      <BillsToPay title="Minhas Contas" />
+    </Module>
   </>
 );
